requirements.txt
absl-py==1.4.0
ale-py==0.8.1
alembic==1.13.1
androguard==4.0.1
anyio==4.1.0
apk-signer==1.1.2
apkInspector==1.2.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
audioread==3.0.1
Babel==2.14.0
backcall==0.2.0
banal==1.0.6
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.0
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.28
dataset==1.6.2
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.4
etils==1.10.0
exceptiongroup==1.2.0
executing==2.0.1
exif==1.3.5
Farama-Notifications==0.0.4
fasteners==0.19
fastjsonschema==2.16.2
Flask==3.0.0
flatbuffers==24.3.25
fonttools==4.39.0
fqdn==1.5.1
frida==16.1.10
fsspec==2024.10.0
gast==0.4.0
glfw==2.7.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
gym==0.25.2
gym-notices==0.0.8
gymnasium==1.0.0
h5py==3.12.1
html5lib==1.1
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib_resources==6.4.5
ipykernel==6.27.1
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.0.2
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.33
jaxlib==0.4.33
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.25.2
keras==3.6.0
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.42.0
loguru==0.7.2
lxml==4.9.4
Mako==1.3.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
ml-dtypes==0.4.1
moviepy==1.0.3
mpi4py==4.0.1
msgpack==1.0.8
mujoco==3.2.5
mujoco-py==2.1.2.14
mutf8==1.0.6
namex==0.0.8
nbclient==0.6.8
nbconvert==7.1.0
nbformat==5.6.1
nest-asyncio==1.5.8
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.59.1
numpy==2.1.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.13.0
overrides==7.4.0
packaging==23.2
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.1.0
plum-py==0.8.2
pooch==1.8.1
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
pydub==0.25.1
pygame==2.1.0
pyglet==1.5.11
Pygments==2.17.2
PyOpenGL==3.1.7
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.3.2
QtPy==2.2.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.13.2
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
Send2Trash==1.8.2
Shimmy==2.0.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.7
SQLAlchemy==1.4.50
stable-baselines==2.10.2
stack-data==0.6.3
tensorboard==2.18.0
tensorboard-data-server==0.7.1
tensorflow==2.18.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.13.0
termcolor==2.3.0
terminado==0.16.0
threadpoolctl==3.2.0
tifffile==2023.12.9
tinycss2==1.1.1
tomli==2.0.1
tornado==6.4
tqdm==4.65.0
traitlets==5.14.0
types-python-dateutil==2.8.19.14
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
widgetsnbextension==4.0.3
wrapt==1.15.0
zipp==3.20.2
